subscriber test working
